fix: move verifier to smt package
